+++ ui/src/main/java/com/example/smonkey_android/feature/friend/RequestData.kt
@@ -3,7 +3,7 @@ package com.example.smonkey_android.feature.friend
 import com.example.data.remote.response.friend.RequestListResponse
 
 data class RequestData(
-    val friendId: Int,
+    val friendId: Long,
     val userName: String,
     val smonkeyName: String,
     val backgroundColor: String,
